WCSS as Rcpp
