+++ src/adjClustFast.cpp
@@ -303,5 +303,55 @@ arma::mat matR_full_rowCumsums(const arma::mat & Csq, const int & h) {
 
 
 
+using namespace Rcpp;
+
+// [[Rcpp::export]]
+NumericVector wcss_single(const arma::SpMat<double> & C, const NumericVector & cluster){
+
+	NumericVector result(max(cluster));
+
+	int baseline = 0;
+
+	// for each feature 
+	for(int i=1; i<cluster.length(); i++){
+
+		// if current and previous features belong to different features
+		if( (cluster(i) != cluster(i-1)) | (i+1 == cluster.length()) ){
+
+			// get submatrix corrsponding to this cluster
+			arma::SpMat<double> C_sub = C.submat(baseline, baseline, i, i);
+
+			// Evalute sum of all matrix elements
+			arma::sp_mat::iterator start = C_sub.begin();
+			arma::sp_mat::iterator end = C_sub.end();
+
+			double total = 0;
+			for(arma::sp_mat::iterator it = start; it != end; ++it){
+				total += (*it);
+			}
+			result[cluster(i-1)-1] = total;
+			baseline = i;
+		}
+	}
+
+	return result;
+}
+    
+
+// [[Rcpp::export]]
+NumericVector WCSS(const arma::SpMat<double> & C, const NumericMatrix & clusterMat){
+
+	NumericVector result(clusterMat.ncol());
+
+	#pragma omp parallel for if(parallelism_enabled)  
+	for(int j=0; j<clusterMat.ncol(); j++){
+
+		NumericVector res = wcss_single(C, clusterMat(_,j) );
+		result[j] = sum(res);
+	}
+
+	return result;
+}
+
